Refactor logging helpers
