+++ src/include/test/unit_logging_time.cc
@@ -1,5 +1,5 @@
 /**
- * @file   unit_logging.cc
+ * @file unit_logging_time.cc
  *
  * @section LICENSE
  *
@@ -33,7 +33,7 @@
 
 #include <catch2/catch_all.hpp>
 #include <thread>
-#include "utils/logging.h"
+#include "utils/logging_time.h"
 
 bool debug = false;
 
@@ -42,25 +42,31 @@ using namespace std::literals::chrono_literals;
 auto duration = 500ms;
 
 TEST_CASE("test", "[logging][log_timer]") {
+  std::cout << "log_timer a(test); -------------------" << std::endl;
   log_timer a("test");
 
   std::this_thread::sleep_for(500ms);
 
+  std::cout << "a.stop() -------------------" << std::endl;
   a.stop();
 
   auto f = _timing_data.get_entries_summed("test");
+  std::cout << f << std::endl;
   CHECK((f <= 520 && f >= 500));
-
+  std::cout << "a.start() -------------------" << std::endl;
   a.start();
   std::this_thread::sleep_for(500ms);
+  std::cout << "a.stop() -------------------" << std::endl;
   a.stop();
 
+  std::cout << "-------------------" << std::endl;
   f = _timing_data.get_entries_summed("test");
+  std::cout << f << std::endl;
   CHECK((f <= 1040 && f >= 1000));
 }
 
 TEST_CASE("noisy test", "[logging][log_timer]") {
-  log_timer a("noisy_test", true);
+  log_timer a("noisy_test");
 
   std::this_thread::sleep_for(500ms);
 
@@ -78,7 +84,7 @@ TEST_CASE("noisy test", "[logging][log_timer]") {
 }
 
 TEST_CASE("interval test", "[logging][log_timer]") {
-  log_timer a("interval_test", true);
+  log_timer a("interval_test");
 
   std::this_thread::sleep_for(500ms);
 
@@ -118,17 +124,6 @@ TEST_CASE("interval test", "[logging][log_timer]") {
 
   f = _timing_data.get_entries_summed("interval_test");
   CHECK((f <= 1040 && f >= 1000));
-}
-
-TEST_CASE("scoped_timer start test", "[logging][scoped_timer]") {
-  scoped_timer a("life_test");
-  std::this_thread::sleep_for(300ms);
-}
-
-TEST_CASE("scoped_timer stop test", "[logging][timing_data]") {
-  std::this_thread::sleep_for(500ms);
-  auto f = _timing_data.get_entries_summed("life_test");
-  CHECK((f <= 320 && f >= 300));
 }
 
 TEST_CASE("ordering", "[logging][log_timer]") {
@@ -173,10 +168,8 @@ TEST_CASE("ordering", "[logging][log_timer]") {
   CHECK((h_t > 499 && h_t < 560));
   CHECK((g_t > 499 && g_t < 560));
   CHECK((f_t > 499 && f_t < 560));
-}
 
-TEST_CASE("memory", "[logging]") {
-  _memory_data.insert_entry(tdb_func__, 8675309);
+  std::cout << _timing_data.dump();
 }
 
 TEST_CASE("multithreaded timing test", "[logging][log_timer]") {
@@ -199,48 +192,6 @@ TEST_CASE("multithreaded timing test", "[logging][log_timer]") {
 
   CHECK(f1 >= 500);
   CHECK(f2 >= 500);
-}
-
-TEST_CASE("multithreaded memory test", "[logging][memory_data]") {
-  auto thread_func = [](const std::string& mem_name, size_t mem_use) {
-    for (int i = 0; i < 10; ++i) {
-      _memory_data.insert_entry(mem_name, mem_use);
-      std::this_thread::sleep_for(10ms);
-    }
-  };
-
-  std::thread t1(thread_func, "multithreaded_memory_test1", 1024);
-  std::thread t2(thread_func, "multithreaded_memory_test2", 2048);
-
-  t1.join();
-  t2.join();
-
-  auto m1 = _memory_data.get_entries_summed("multithreaded_memory_test1");
-  auto m2 = _memory_data.get_entries_summed("multithreaded_memory_test2");
-
-  CHECK(m1 >= 10 * 1024 / (1024 * 1024));
-  CHECK(m2 >= 10 * 2048 / (1024 * 1024));
-}
-
-TEST_CASE("multithreaded count test", "[logging][count_data]") {
-  auto thread_func = [](const std::string& count_name, size_t count) {
-    for (int i = 0; i < 10; ++i) {
-      _count_data.insert_entry(count_name, count);
-      std::this_thread::sleep_for(10ms);
-    }
-  };
-
-  std::thread t1(thread_func, "multithreaded_count_test1", 1);
-  std::thread t2(thread_func, "multithreaded_count_test2", 2);
-
-  t1.join();
-  t2.join();
-
-  auto c1 = _count_data.get_entries_summed("multithreaded_count_test1");
-  auto c2 = _count_data.get_entries_summed("multithreaded_count_test2");
-
-  CHECK(c1 == 10);
-  CHECK(c2 == 20);
 }
 
 TEST_CASE("highly concurrent timing test", "[logging][log_timer]") {
@@ -267,69 +218,4 @@ TEST_CASE("highly concurrent timing test", "[logging][log_timer]") {
 
   auto f = _timing_data.get_entries_summed(timer_name);
   CHECK(f >= num_threads * num_iterations * 1);
-}
-
-TEST_CASE("highly concurrent memory test", "[logging][memory_data]") {
-  constexpr auto timer_name = "highly_concurrent_memory_test";
-  constexpr int num_iterations = 100;
-  auto thread_func = []() {
-    for (int i = 0; i < num_iterations; ++i) {
-      memory_data::memory_type mem_usage = 1024;  // Simulating memory usage
-      _memory_data.insert_entry(timer_name, mem_usage);
-      std::this_thread::sleep_for(1ms);
-    }
-  };
-
-  auto num_threads = 20;
-  std::vector<std::thread> threads;
-  threads.reserve(num_threads);
-  for (int i = 0; i < num_threads; ++i) {
-    threads.emplace_back(thread_func);
-  }
-
-  for (auto& thread : threads) {
-    thread.join();
-  }
-
-  auto f = _memory_data.get_entries_summed(timer_name);
-  CHECK(
-      f >=
-      num_threads * num_iterations * 1024 / (1024 * 1024));  // Convert to MiB
-}
-
-TEST_CASE("highly concurrent count test", "[logging][count_data]") {
-  constexpr auto timer_name = "highly_concurrent_count_test";
-  constexpr int num_iterations = 100;
-  auto thread_func = []() {
-    for (int i = 0; i < num_iterations; ++i) {
-      _count_data.insert_entry(timer_name, 1);
-      std::this_thread::sleep_for(1ms);
-    }
-  };
-
-  auto num_threads = 20;
-  std::vector<std::thread> threads;
-  threads.reserve(num_threads);
-  for (int i = 0; i < num_threads; ++i) {
-    threads.emplace_back(thread_func);
-  }
-
-  for (auto& thread : threads) {
-    thread.join();
-  }
-
-  auto f = _count_data.get_entries_summed(timer_name);
-  CHECK(f >= num_threads * num_iterations);
-}
-
-TEST_CASE("dump", "[logging]") {
-  _timing_data.insert_entry(
-      "a",
-      std::chrono::high_resolution_clock::now() -
-          std::chrono::high_resolution_clock::now());
-  _count_data.insert_entry("b", 1);
-  _memory_data.insert_entry("c", 2);
-  std::cout << _timing_data.dump();
-  std::cout << _count_data.dump();
-  std::cout << _memory_data.dump();
 }
