+++ src/include/utils/logging_scoped_time.h
@@ -1,0 +1,67 @@
+/**
+ * @file   logging.h
+ *
+ * @section LICENSE
+ *
+ * The MIT License
+ *
+ * @copyright Copyright (c) 2023 TileDB, Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ *
+ * @section DESCRIPTION
+ *
+ */
+
+#ifndef TDB_LOGGING_H
+#define TDB_LOGGING_H
+
+#include <chrono>
+#include <iomanip>
+#include <iostream>
+#include <map>
+#include <mutex>
+#include <set>
+#include <string>
+#include <vector>
+#include "utils/logging_time.h"
+
+/**
+ * Scoped timer class for logging timing data.  Internnally maintains a start
+ * time and a stop time (which are std::chrono::time_points).  The constructor
+ * and destructor control operation and logging of the timer. It inherits from
+ * `log_timer` but invokes `stop()` in its destructor. It is intended to measure
+ * the lifetime of a scope.  It begins timing when it is constructed and stops
+ * timing when it is destructed.
+ */
+class scoped_timer : public log_timer {
+ public:
+  explicit scoped_timer(const std::string& msg = "", bool noisy = false)
+      : log_timer(msg, noisy) {
+  }
+
+  /**
+   * Stop the timer when it goes out of scope.
+   */
+  ~scoped_timer() {
+    this->stop();
+  }
+};
+
+#endif  // TDB_LOGGING_H
