+++ src/include/utils/logging_time.h
@@ -367,380 +367,4 @@ class log_timer {
   }
 };
 
-/**
- * Scoped timer class for logging timing data.  Internnally maintains a start
- * time and a stop time (which are std::chrono::time_points).  The constructor
- * and destructor control operation and logging of the timer. It inherits from
- * `log_timer` but invokes `stop()` in its destructor. It is intended to measure
- * the lifetime of a scope.  It begins timing when it is constructed and stops
- * timing when it is destructed.
- */
-class scoped_timer : public log_timer {
- public:
-  explicit scoped_timer(const std::string& msg = "", bool noisy = false)
-      : log_timer(msg, noisy) {
-  }
-
-  /**
-   * Stop the timer when it goes out of scope.
-   */
-  ~scoped_timer() {
-    this->stop();
-  }
-};
-
-/**
- * Singleton class for recording memory consumption.  Internally maintains a
- * multimap of names and memory consumption values.  The `insert_entry()` method
- * is used to record memory consumption.  The `get_entries_*()` methods are used
- * to query the recorded memory consumption.
- *
- * Data are stored as bytes.  The `get_entries_*()` methods return values
- * as MiB (2^10 bytes).
- */
-class memory_data {
- public:
-  using memory_type = size_t;
-
- private:
-  std::multimap<std::string, memory_type> memory_usages_;
-  mutable std::mutex mtx_;
-  bool verbose_{false};
-
-  /**
-   * Constructor.  Private to enforce singleton pattern.
-   */
-  memory_data() = default;
-  ~memory_data() = default;
-
- public:
-  /**
-   * Copy constructor.  Deleted to enforce singleton pattern.
-   */
-  memory_data(const memory_data&) = delete;
-  memory_data& operator=(const memory_data&) = delete;
-
-  /**
-   * Get a reference to the singleton instance of the class.
-   * @return Reference to the singleton instance of the class.
-   */
-  static memory_data& get_instance() {
-    static std::once_flag flag;
-    // This will leak, but it's okay - it's the Trusty Leaky Singleton pattern.
-    static memory_data* instance;
-    std::call_once(flag, []() { instance = new memory_data(); });
-    return *instance;
-  }
-
-  /**
-   * Insert a memory consumption entry into the multimap.
-   * @param name The name to be associated with the memory consumption.
-   * @param use The memory consumption to be recorded (in bytes).
-   */
-  void insert_entry(const std::string& name, const memory_type& use) {
-    std::lock_guard<std::mutex> lock(mtx_);
-    memory_usages_.insert(std::make_pair(name, use));
-  }
-
-  /**
-   * Get the memory consumption entries associated with a name.
-   * @param string Name to be queried.
-   * @return Vector of memory consumption values associated with the name.
-   */
-  auto get_entries_separately(const std::string& string) const {
-    std::lock_guard<std::mutex> lock(mtx_);
-    std::vector<double> usages;
-
-    auto range = memory_usages_.equal_range(string);
-    for (auto i = range.first; i != range.second; ++i) {
-      usages.push_back(i->second / (1024 * 1024));
-    }
-    return usages;
-  }
-
-  /**
-   * Get the sum of the memory consumption entries associated with a name.
-   * @param string Name to be queried.
-   * @return Vector of memory consumption values associated with the name.
-   */
-  auto get_entries_summed(const std::string& string) const {
-    std::lock_guard<std::mutex> lock(mtx_);
-    double sum = 0.0;
-    auto range = memory_usages_.equal_range(string);
-    for (auto i = range.first; i != range.second; ++i) {
-      sum += i->second;
-    }
-    return sum / (1024 * 1024);
-  }
-
-  /**
-   * Get the names associated with the memory consumption entries.
-   * @return Vector of names associated with the memory consumption entries.
-   */
-  auto get_usage_names() const {
-    std::lock_guard<std::mutex> lock(mtx_);
-    std::set<std::string> multinames;
-
-    std::vector<std::string> names;
-
-    for (const auto& i : memory_usages_) {
-      multinames.insert(i.first);
-    }
-    for (const auto& i : multinames) {
-      names.push_back(i);
-    }
-    return names;
-  }
-
-  void set_verbose(bool verbose) {
-    verbose_ = verbose;
-  }
-
-  bool get_verbose() const {
-    return verbose_;
-  }
-
-  std::string dump() const {
-    std::ostringstream oss;
-    auto usage_names = get_usage_names();
-    for (const auto& name : usage_names) {
-      auto usages = get_entries_separately(name);
-      double sum_mib = get_entries_summed(name);
-      double average_mib = sum_mib / usages.size();
-
-      oss << name << ":";
-      oss << " count: " << usages.size() << ",";
-      oss << " sum: " << format_memory_mib(sum_mib) << ",";
-      oss << " avg: " << format_memory_mib(average_mib);
-      oss << "\n";
-    }
-    return oss.str();
-  }
-
- private:
-  std::string format_memory_mib(double memory_mib) const {
-    std::ostringstream oss;
-    oss << std::fixed << std::setprecision(2);
-
-    if (memory_mib < 1) {
-      oss << (memory_mib * 1024);
-      return oss.str() + " KiB";
-    } else if (memory_mib < 1024) {
-      oss << memory_mib;
-      return oss.str() + " MiB";
-    } else {
-      oss << (memory_mib / 1024);
-      return oss.str() + " GiB";
-    }
-  }
-};
-
-inline memory_data& get_memory_data_instance() {
-  return memory_data::get_instance();
-}
-
-static memory_data& _memory_data{get_memory_data_instance()};
-
-/**
- * Singleton class for recording miscellaneous counts.  Internally maintains a
- * multimap of names and count values.  The `insert_entry()` method
- * is used to record the count.  The `get_entries_*()` methods are used
- * to query the recorded counts.
- *
- * Data are stored as integral values.  The `get_entries_*()` methods return
- * integral values.
- */
-class count_data {
- public:
-  using count_type = size_t;
-
- private:
-  std::multimap<std::string, count_type> count_usages_;
-  mutable std::mutex mtx_;
-  bool verbose_{false};
-
-  /**
-   * Constructor.  Private to enforce singleton pattern.
-   */
-  count_data() = default;
-  ~count_data() = default;
-
- public:
-  /**
-   * Copy constructor.  Deleted to enforce singleton pattern.
-   */
-  count_data(const count_data&) = delete;
-  count_data& operator=(const count_data&) = delete;
-
-  /**
-   * Get a reference to the singleton instance of the class.
-   * @return Reference to the singleton instance of the class.
-   */
-  static count_data& get_instance() {
-    static std::once_flag flag;
-    // This will leak, but it's okay - it's the Trusty Leaky Singleton pattern.
-    static count_data* instance;
-    std::call_once(flag, []() { instance = new count_data(); });
-    return *instance;
-  }
-
-  /**
-   * Insert a count entry into the multimap.
-   * @param name The name to be associated with the count.
-   * @param use The count to be recorded (in bytes).
-   */
-  void insert_entry(const std::string& name, const count_type& use) {
-    std::lock_guard<std::mutex> lock(mtx_);
-    count_usages_.insert(std::make_pair(name, use));
-  }
-
-  /**
-   * Get the count entries associated with a name.
-   * @param string Name to be queried.
-   * @return Vector of count values associated with the name.
-   */
-  auto get_entries_separately(const std::string& string) const {
-    std::lock_guard<std::mutex> lock(mtx_);
-    std::vector<double> usages;
-
-    auto range = count_usages_.equal_range(string);
-    for (auto i = range.first; i != range.second; ++i) {
-      usages.push_back(i->second);
-    }
-    return usages;
-  }
-
-  /**
-   * Get the sum of the count entries associated with a name.
-   * @param string Name to be queried.
-   * @return Vector of count values associated with the name.
-   */
-  auto get_entries_summed(const std::string& string) const {
-    std::lock_guard<std::mutex> lock(mtx_);
-    double sum = 0.0;
-    auto range = count_usages_.equal_range(string);
-    for (auto i = range.first; i != range.second; ++i) {
-      sum += i->second;
-    }
-    return sum;
-  }
-
-  /**
-   * Get the names associated with the count entries.
-   * @return Vector of names associated with the count entries.
-   */
-  std::vector<std::string> get_usage_names() const {
-    std::lock_guard<std::mutex> lock(mtx_);
-    std::set<std::string> multinames;
-
-    std::vector<std::string> names;
-
-    for (const auto& i : count_usages_) {
-      multinames.insert(i.first);
-    }
-    for (const auto& i : multinames) {
-      names.push_back(i);
-    }
-    return names;
-  }
-
-  void set_verbose(bool verbose) {
-    verbose_ = verbose;
-  }
-
-  bool get_verbose() const {
-    return verbose_;
-  }
-
-  std::string dump() const {
-    std::ostringstream oss;
-    auto usage_names = get_usage_names();
-    for (const auto& name : usage_names) {
-      auto usages = get_entries_separately(name);
-      double sum = get_entries_summed(name);
-      double average = sum / usages.size();
-
-      oss << name << ":";
-      oss << " count: " << usages.size() << ",";
-      oss << " sum: " << sum << ",";
-      oss << " avg: " << average;
-      oss << "\n";
-    }
-    return oss.str();
-  }
-};
-
-inline count_data& get_count_data_instance() {
-  return count_data::get_instance();
-}
-
-static count_data& _count_data{get_count_data_instance()};
-
-#if 0
-class stats_data {
-  public:
-    enum operation_type {open_array, submit_query};
-
-    struct stats_type {
-      operation_type operation;
-      std::string uri;
-      std::string location;
-    };
-
-    using stats_map = std::multimap<stats_type, std::string>;
-
-  private:
-
-    stats_map stats_;
-    bool verbose_{false};
-
-    /**
-     * Constructor.  Private to enforce singleton pattern.
-     */
-    stats_data() = default;
-    ~stats_data() = default;
-
-  public:
-
-    /**
-     * Copy constructor.  Deleted to enforce singleton pattern.
-     */
-    stats_data(const stats_data&) = delete;
-    stats_data& operator=(const stats_data&) = delete;
-
-    /**
-     * Get a reference to the singleton instance of the class.
-     * @return Reference to the singleton instance of the class.
-     */
-    static stats_data& get_instance() {
-      static stats_data instance;
-      return instance;
-    }
-
-    /**
-     * Insert a memory consumption entry into the multimap.
-     * @param name The name to be associated with the memory consumption.
-     * @param use The memory consumption to be recorded (in bytes).
-     */
-    void insert_entry(const stats_type& key, const std::string& stats) {
-      stats_.insert({key, stats});
-    }
-
-    /**
-     * Get the memory consumption entries associated with a name.
-     * @param string Name to be queried.
-     * @return Vector of memory consumption values associated with the name.
-     */
-    auto get_entries_separately(const std::string& string) const {
-      std::vector<double> usages;
-
-      auto range = stats_.equal_range(string);
-      for (auto i = range.first; i != range.second; ++i) {
-        usages.push_back(i->second / (1024*1024));
-      }
-      return usages;
-    }
-};
-#endif
-
 #endif  // TDB_LOGGING_H
